refactor/test: migration to vitest
