+++ src/core.js
@@ -12,6 +12,5 @@ export {
     useId
 } from "./hooks/create-hooks.js";
 export { render, h, Mark, Fragment, h as createElement } from "./render.js";
-export { template } from "./template.js";
 export { css } from "./css.js";
 export { useContext, createContext, useProvider } from "./context.js";
