+++ tsconfig.json
@@ -6,7 +6,7 @@
         "noUnusedLocals": true,
         "jsx": "react-jsx",
         "checkJs": true,
-        "jsxImportSource": "../",
+        "jsxImportSource": "atomico",
         "target": "ESNext",
         "module": "NodeNext",
         "moduleResolution": "NodeNext",
@@ -16,6 +16,7 @@
         "noEmit": true,
         "lib": ["ESNext", "DOM", "DOM.Iterable"],
         "paths": {
+            "atomico": ["./"],
             "core": ["./types/core.d.ts"],
             "hooks": ["./types/hooks.d.ts"],
             "dom": ["./types/dom.d.ts"],
